pyproject.toml
[project]
name = "movie-recommendation-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "nltk>=3.9.4",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.2",
    "flask>=3.0.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.1",
]
